Create main dir and move forms into it
